Allow user to manually define a project list
Sometimes, the cyclops server might not have access to the sentry SQL
server where the project public secret keys are stored. In that case,
the user will want to manually define the list of project keys for which
the cyclops can act as a proxy. We defined the PROJECT_KEYS
configuration variable for that purpose.